+++ src/Contracts/PaperModel.php
@@ -6,6 +6,7 @@ namespace JacobJoergensen\LaravelPaper\Contracts;
 
 use Illuminate\Database\Eloquent\Model;
 use JacobJoergensen\LaravelPaper\PaperQueryBuilder;
+use JacobJoergensen\LaravelPaper\Relations\PaperRelation;
 
 /**
  * @phpstan-require-extends Model
@@ -16,6 +17,11 @@ interface PaperModel
      * @return PaperQueryBuilder<static&Model>
      */
     public static function query(): PaperQueryBuilder;
+
+    /**
+     * @return array<string, PaperRelation<Model&PaperModel>>
+     */
+    public function paperRelations(): array;
 
     public function getContentPath(): string;